+++ lib/navbar_router.dart
@@ -2,3 +2,4 @@ library navbar_router;
 
 export 'src/navbar_router.dart';
 export 'src/navbar_notifier.dart';
+export 'src/navbar_decoration.dart';
