+++ lib/src/navbar_decoration.dart
@@ -1,0 +1,81 @@
+import 'package:flutter/material.dart';
+
+class NavbarItem {
+  const NavbarItem(this.iconData, this.text, {this.backgroundColor});
+
+  /// IconData for the navbar item
+  final IconData iconData;
+
+  /// label for the navbar item
+  final String text;
+
+  /// background color for the navbar item whnen type is [NavbarType.shifting]
+  final Color? backgroundColor;
+}
+
+class NavbarDecoration {
+  /// The type of the Navbar to be displayed
+  /// [BottomNavigationBarType.fixed] or [BottomNavigationBarType.shifting]
+  final BottomNavigationBarType? navbarType;
+
+  /// The backgroundColor of the Navbar
+  final Color? backgroundColor;
+
+  /// Defines whether the Navbar is extended in Desktop mode
+  /// defaultst to false
+  final bool isExtended;
+
+  /// The color of the unselected item
+  final Color? unselectedItemColor;
+
+  /// The elevation shadown on the edges of bottomnavigationbar
+  final double? elevation;
+
+  /// The color of the unselected item icon
+  final Color? unselectedIconColor;
+
+  /// Whether or not to show the unselected label text
+  final bool? showUnselectedLabels;
+
+  /// The color of the unselected label text
+  final Color? unselectedLabelColor;
+
+  /// The color of the label text
+  final Color? selectedLabelColor;
+
+  /// whether or not to show the selected label text
+  final bool? showSelectedLabels;
+
+  /// haptic feedbakc when the item is selected
+  final bool? enableFeedback;
+
+  /// the text style of the selected label
+  final TextStyle? selectedLabelTextStyle;
+
+  /// the text style of the unselected labels
+  final TextStyle? unselectedLabelTextStyle;
+
+  /// iconTheme for the selected icon
+  final IconThemeData? selectedIconTheme;
+
+  /// iconTheme for the unselected icon
+  final IconThemeData? unselectedIconTheme;
+
+  NavbarDecoration({
+    this.backgroundColor,
+    this.elevation,
+    this.enableFeedback,
+    this.isExtended = false,
+    this.navbarType,
+    this.showSelectedLabels,
+    this.showUnselectedLabels = true,
+    this.selectedIconTheme,
+    this.selectedLabelColor,
+    this.selectedLabelTextStyle,
+    this.unselectedIconTheme,
+    this.unselectedLabelTextStyle,
+    this.unselectedIconColor,
+    this.unselectedItemColor,
+    this.unselectedLabelColor,
+  });
+}
